refactor: Injector 제거
